Added Airtable integration to util function
