--- processes/util/airtable_authentication.py
@@ -1,8 +1,0 @@
-import os
-
-def create_authorization_header():
-    bearer_token = os.environ.get("AIRTABLE_KEY")
-
-    headers = {"Authorization": f"Bearer {bearer_token}"}
-
-    return headers
